requirements.txt
fastapi==0.115.5
pydantic-settings==2.6.1
uvicorn==0.32.1
python-socketio==5.11.4
jinja2==3.1.4
websockets==14.1
redis==5.2.0
pymongo==4.10.1
email-validator==2.2.0